design: Navbar 디자인 수정
