+++ src/Components/Navbar/BasicNav.jsx
@@ -1,32 +1,26 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import IconArrowLeft from "./icon/IconArrowLeft";
-import IconMoreVertical from "./icon/IconMoreVertical";
+import IconArrowLeft from "../icon/IconArrowLeft";
+import IconMoreVertical from "../icon/IconMoreVertical";
 
 const Container = styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
   position: fixed;
+  top: 0;
   width: 100vw;
   height: 48px;
   padding: 0 12px 0 16px;
+  border-bottom: 0.5px solid #dbdbdb;
+  background-color: #ffffff;
 `;
 
 const Back = styled(Link)`
   width: 22px;
   height: 22px;
   cursor: pointer;
-`;
-
-const SellerName = styled.h2`
-  flex: 1;
-  margin-left: 10px;
-  font-family: "Godo";
-  font-size: 1.4rem;
-  font-weight: 500;
-  line-height: 1.753rem;
 `;
 
 const More = styled.button`
@@ -47,18 +41,13 @@ const MoreIcon = styled(IconMoreVertical)`
   width: 24px;
   height: 24px;
 `;
-/**
- *
- * @param {{sellerName: string}} param0
- * @returns
- */
-function ChatNav({ sellerName }) {
+
+function BasicNav() {
   return (
     <Container>
       <Back>
         <BackBtnIcon />
       </Back>
-      <SellerName>{sellerName}</SellerName>
       <More type="button">
         <MoreIcon />
       </More>
@@ -66,4 +55,4 @@ function ChatNav({ sellerName }) {
   );
 }
 
-export default ChatNav;
+export default BasicNav;
