+++ src/Components/Navbar/MainNav.jsx
@@ -1,16 +1,19 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import IconSearch from "./icon/IconSearch";
+import IconSearch from "../icon/IconSearch";
 
 const Container = styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
   position: fixed;
+  top: 0;
   width: 100vw;
   height: 48px;
   padding: 0 16px;
+  border-bottom: 0.5px solid #dbdbdb;
+  background-color: #ffffff;
 `;
 
 const Title = styled.h2`
