+++ src/Components/Navbar/SearchNav.jsx
@@ -1,38 +1,56 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import CommonButton from "./button/CommonButton";
-import IconArrowLeft from "./icon/IconArrowLeft";
+import IconArrowLeft from "../icon/IconArrowLeft";
 
 const Container = styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
   position: fixed;
+  top: 0;
   width: 100vw;
   height: 48px;
   padding: 0 16px;
+  border-bottom: 0.5px solid #dbdbdb;
+  background-color: #ffffff;
 `;
-
 const Back = styled(Link)`
   width: 22px;
   height: 22px;
   cursor: pointer;
+`;
+
+const Input = styled.input`
+  flex-grow: 1;
+  height: 32px;
+  margin-left: 20px;
+  padding: 7px 16px;
+  border: none;
+  border-radius: 32px;
+  background-color: #f2f2f2;
+  font-size: 1.4rem;
+  font-family: "LINESeedKR-Bd";
+
+  ::placeholder {
+    color: #c4c4c4;
+  }
+  :focus {
+    outline: none;
+  }
 `;
 const BackBtnIcon = styled(IconArrowLeft)`
   width: 22px;
   height: 22px;
 `;
 
-function UploadNav() {
+export default function SearchNav() {
   return (
     <Container>
       <Back>
         <BackBtnIcon />
       </Back>
-      <CommonButton size="ms" bgColor="accent" children="저장" />
+      <Input placeholder="계정 검색" />
     </Container>
   );
 }
-
-export default UploadNav;
